+++ src/Integrals/SlaterCD.C
@@ -42,7 +42,7 @@
         
  }
  
- double SlaterCD::fk(double a, double ab, int k,int n)
+ double SlaterCD::fk(double a, double ab, size_t k,size_t n)
 {
     assert(n>0);
     assert(k>=0);
@@ -50,51 +50,51 @@
     return qchem::Fact[k]*(n/pow(ab,k+1)+1/pow(a,k+1));
 }
 
-double SlaterCD::Coulomb_R0(int la,int lc) const
+double SlaterCD::Coulomb_R0(size_t la,size_t lc) const
 {
     assert(la<=LMax);
     assert(lc<=LMax);
-    int Lab_p=2*la+3; // first term r_1^2
-    int Lcd_m=2*lc+1; // first term r_2
-    int Lab_m=2*la+1; // second term r_1
-    int Lcd_p=2*lc+3; // second term r_2^2
+    size_t Lab_p=2*la+3; // first term r_1^2
+    size_t Lcd_m=2*lc+1; // first term r_2
+    size_t Lab_m=2*la+1; // second term r_1
+    size_t Lcd_p=2*lc+3; // second term r_2^2
     //cout << la << " " << lc << " " << k << " " << Lab_p << " " << Lcd_p << endl;
     return (Iab(Lab_m,Lcd_p)+Icd(Lcd_m,Lab_p));
 }
 
-Vector<double> SlaterCD::Coulomb_Rk(int la,int lc) const
+Vector<double> SlaterCD::Coulomb_Rk(size_t la,size_t lc) const
 {
     assert(la<=LMax);
     assert(lc<=LMax);
     Vector<double> ret(la+lc+1,0.0);
-    int i=1;
-    for (int k=0;k<=2*std::min(la,lc);k+=2)
+    size_t i=1;
+    for (size_t k=0;k<=2*std::min(la,lc);k+=2)
     {
-        int Lab_p=2*la+3+k; // first term r_1^2
-        int Lcd_m=2*lc+1-k; // first term r_2
-        int Lab_m=2*la+1-k; // second term r_1
-        int Lcd_p=2*lc+3+k; // second term r_2^2
+        size_t Lab_p=2*la+3+k; // first term r_1^2
+        size_t Lcd_m=2*lc+1-k; // first term r_2
+        size_t Lab_m=2*la+1-k; // second term r_1
+        size_t Lcd_p=2*lc+3+k; // second term r_2^2
         //cout << la << " " << lc << " " << k << " " << Lab_p << " " << Lcd_p << endl;
         ret(i++)=(Iab(Lab_m,Lcd_p)+Icd(Lcd_m,Lab_p));
     }
     return ret;
 }
 
-Vector<double> SlaterCD::ExchangeRk(int la,int lb) const
+Vector<double> SlaterCD::ExchangeRk(size_t la,size_t lb) const
 {
     assert(la<=LMax);
     assert(lb<=LMax);
-    int kmin=std::abs(la-lb);
-    int kmax=la+lb;
-    int N=(kmax-kmin)/2+1;
+    size_t kmin=std::abs((int)la-(int)lb);
+    size_t kmax=la+lb;
+    size_t N=(kmax-kmin)/2+1;
     Vector<double> ret(N,0.0);
-    int i=1;
-    for (int k=kmin;k<=kmax;k+=2)
+    size_t i=1;
+    for (size_t k=kmin;k<=kmax;k+=2)
     {
-        int Lab_p=la+lb+3+k; // first term r_1^2
-        int Lcd_m=la+lb+1-k; // first term r_2
-        int Lab_m=la+lb+1-k; // second term r_1
-        int Lcd_p=la+lb+3+k; 
+        size_t Lab_p=la+lb+3+k; // first term r_1^2
+        size_t Lcd_m=la+lb+1-k; // first term r_2
+        size_t Lab_m=la+lb+1-k; // second term r_1
+        size_t Lcd_p=la+lb+3+k; 
         ret(i++)=(Iab(Lab_m,Lcd_p)+Icd(Lcd_m,Lab_p)); //(2*k+1)???
     }
     return ret;
